\name{LazyCD-package}
\alias{LazyCD-package}
\alias{LazyCD}
\docType{package}
\title{lazyCD package}
\packageTitle{LazyCD}
}
\description{
\packageDescription{LazyCD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LazyCD}
\packageIndices{LazyCD}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{LazyCD}

Maintainer: \packageMaintainer{LazyCD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
